+++ UnitTests/Lemoine.UnitTests/WithHourTimeStamp.cs
@@ -1,4 +1,5 @@
 // Copyright (C) 2009-2023 Lemoine Automation Technologies
+// Copyright (C) 2025 Atsora Solutions
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -15,9 +16,6 @@ namespace Lemoine.UnitTests
   {
     static readonly ILog log = LogManager.GetLogger(typeof (WithHourTimeStamp).FullName);
 
-    #region Getters / Setters
-    #endregion // Getters / Setters
-
     #region Constructors
     /// <summary>
     /// Description of the constructor
@@ -29,7 +27,7 @@ namespace Lemoine.UnitTests
     #endregion // Constructors
 
     #region Methods
-    protected override DateTime T(int n)
+    protected override DateTime T(double n)
     {
       return m_baseDateTime.AddHours (n);
     }
