+++ UnitTests/Lemoine.UnitTests/WithMinuteTimeStamp.cs
@@ -1,4 +1,5 @@
 // Copyright (C) 2009-2023 Lemoine Automation Technologies
+// Copyright (C) 2025 Atsora Solutions
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -14,11 +15,6 @@ namespace Lemoine.UnitTests
   public class WithMinuteTimeStamp: WithTimeStamp
   {
     static readonly ILog log = LogManager.GetLogger(typeof (WithMinuteTimeStamp).FullName);
-
-    #region Getters / Setters
-    #endregion // Getters / Setters
-
-    #region Constructors
     /// <summary>
     /// Description of the constructor
     /// </summary>
@@ -26,10 +22,9 @@ namespace Lemoine.UnitTests
       : base (baseDateTime)
     {
     }
-    #endregion // Constructors
 
     #region Methods
-    protected override DateTime T(int n)
+    protected override DateTime T(double n)
     {
       return m_baseDateTime.AddMinutes (n);
     }
